backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# LangChain core + community + Groq integration
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.1.0

# Text splitters - using custom lightweight implementation (no transformers/torch dependency)

# Embeddings - using scikit-learn TF-IDF + SVD (torch-free)
scikit-learn>=1.3.0

faiss-cpu>=1.7.4
pypdf>=3.17.4
python-dotenv>=1.0.0
numpy>=1.24.3

